Move code from ion-sfu to pkg/sfu
