+++ pkg/sfu/downtrack.go
@@ -13,7 +13,7 @@ import (
 	"github.com/pion/transport/packetio"
 	"github.com/pion/webrtc/v3"
 
-	"github.com/pion/ion-sfu/pkg/buffer"
+	"github.com/livekit/livekit-server/pkg/sfu/buffer"
 )
 
 // DownTrackType determines the type of track
