+++ pkg/sfu/sequencer.go
@@ -4,7 +4,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/pion/ion-sfu/pkg/buffer"
+	"github.com/livekit/livekit-server/pkg/sfu/buffer"
 )
 
 const (
