+++ pkg/sfu/stats/stream.go
@@ -5,7 +5,7 @@ import (
 	"sync"
 	"sync/atomic"
 
-	"github.com/pion/ion-sfu/pkg/buffer"
+	"github.com/livekit/livekit-server/pkg/sfu/buffer"
 	"github.com/prometheus/client_golang/prometheus"
 )
